menambahkan menu invoice

+++ application/controllers/dashboard.php
@@ -44,10 +44,16 @@ class Dashboard extends CI_controller{
     }
 
     public function proses_pesanan(){
-        $this->cart->destroy();
-        $this->load->view('templates/header');
-        $this->load->view('templates/sidebar');
-        $this->load->view('pages/proses_pesanan');
-        $this->load->view('templates/footer');
+        $is_processed = $this->model_invoice->index();
+            if($is_processed){
+                $this->cart->destroy();
+                $this->load->view('templates/header');
+                $this->load->view('templates/sidebar');
+                $this->load->view('pages/proses_pesanan');
+                $this->load->view('templates/footer');
+            }else{
+                echo "pesanan gagal di proses";
+            }
     }
+        
 }
